Moves base_field_comparator & field_comparator into MultiPredicateJoinEvaluator
