--- src/lib/operators/multi_predicate_join/base_field_comparator.hpp
@@ -1,17 +1,0 @@
-#pragma once
-
-#include "types.hpp"
-
-namespace opossum {
-
-namespace mpj {
-class BaseFieldComparator : public Noncopyable {
- public:
-  virtual bool compare(const RowID& left, const RowID& right) const = 0;
-
-  virtual ~BaseFieldComparator() = default;
-};
-
-}  // namespace mpj
-
-}  // namespace opossum
